+++ src/pyke/endpoints/clash.py
@@ -1,12 +1,11 @@
 from typing import Any
 
+from .._base_riot_client import _BaseRiotClient
 from ..enums.region import Region
-
-from .._base_client import _BaseApiClient
 
 
 class ClashEndpoint:
-    def __init__(self, client: _BaseApiClient):
+    def __init__(self, client: _BaseRiotClient):
         self._client = client
 
     async def by_puuid(self, region: Region, puuid: str) -> list[dict[Any, Any]]:
@@ -24,7 +23,7 @@ class ClashEndpoint:
         """  # fmt: skip
 
         path = f"/lol/clash/v1/players/by-puuid/{puuid}"
-        data = await self._client._region_request(region=region, path=path)
+        data = await self._client._request(region=region, path=path)
 
         return data
 
@@ -43,7 +42,7 @@ class ClashEndpoint:
         """  # fmt: skip
 
         path = f"/lol/clash/v1/teams/{team_id}"
-        data = await self._client._region_request(region=region, path=path)
+        data = await self._client._request(region=region, path=path)
 
         return data
 
@@ -61,7 +60,7 @@ class ClashEndpoint:
         """  # fmt: skip
 
         path = "/lol/clash/v1/tournaments"
-        data = await self._client._region_request(region=region, path=path)
+        data = await self._client._request(region=region, path=path)
 
         return data
 
@@ -82,7 +81,7 @@ class ClashEndpoint:
         """  # fmt: skip
 
         path = f"/lol/clash/v1/tournaments/by-team/{team_id}"
-        data = await self._client._region_request(region=region, path=path)
+        data = await self._client._request(region=region, path=path)
 
         return data
 
@@ -103,6 +102,6 @@ class ClashEndpoint:
         """  # fmt: skip
 
         path = f"/lol/clash/v1/tournaments/{tournament_id}"
-        data = await self._client._region_request(region=region, path=path)
+        data = await self._client._request(region=region, path=path)
 
         return data