+++ src/pyke/endpoints/lol_status.py
@@ -1,12 +1,11 @@
 from typing import Any
 
+from .._base_riot_client import _BaseRiotClient
 from ..enums.region import Region
-
-from .._base_client import _BaseApiClient
 
 
 class StatusEndpoint:
-    def __init__(self, client: _BaseApiClient):
+    def __init__(self, client: _BaseRiotClient):
         self._client = client
 
     async def platform_data(self, region: Region) -> dict[Any, Any]:
@@ -23,6 +22,6 @@ class StatusEndpoint:
         """  # fmt: skip
 
         path = "/lol/status/v4/platform-data"
-        data = await self._client._region_request(region=region, path=path)
+        data = await self._client._request(region=region, path=path)
 
         return data