+++ src/pyke/endpoints/summoner.py
@@ -1,11 +1,11 @@
 from typing import Any
 
-from .._base_client import _BaseApiClient
+from .._base_riot_client import _BaseRiotClient
 from ..enums.region import Region
 
 
 class SummonerEndpoint:
-    def __init__(self, client: _BaseApiClient):
+    def __init__(self, client: _BaseRiotClient):
         self._client = client
 
     async def by_puuid(self, region: Region, puuid: str) -> dict[Any, Any]:
@@ -23,6 +23,6 @@ class SummonerEndpoint:
         """  # fmt: skip
 
         path = f"/lol/summoner/v4/summoners/by-puuid/{puuid}"
-        data = await self._client._region_request(region=region, path=path)
+        data = await self._client._request(region=region, path=path)
 
         return data